refactor: move classes to new domain
