+++ paper/src/main/java/us/crazycrew/crazycrates/paper/support/holograms/DecentHologramsSupport.java
@@ -1,7 +1,7 @@
-package com.badbones69.crazycrates.paper.support.holograms;
+package us.crazycrew.crazycrates.paper.support.holograms;
 
-import com.badbones69.crazycrates.paper.CrazyCrates;
-import com.badbones69.crazycrates.paper.Methods;
+import us.crazycrew.crazycrates.paper.CrazyCrates;
+import us.crazycrew.crazycrates.paper.support.Methods;
 import com.badbones69.crazycrates.paper.api.interfaces.HologramController;
 import com.badbones69.crazycrates.paper.api.objects.Crate;
 import org.bukkit.plugin.java.JavaPlugin;
